now compiles with c++11 on mac os x

+++ src/Cell/Cell.cpp
@@ -179,7 +179,7 @@ vector<vector<double> >* Cell::getNeighbors(boost::shared_ptr<Sublattice>& sl1, 
     return &neighborCache[name];
 }
 
-int Cell::size()
+size_t Cell::size()
 {
     return sublatticeInfo.size();
 }
